Web 3D Demo
Added content of the 'dist' directory from my Web 3D Demo repo (https://github.com/samstalhandske/web_3d_demo) to my website.

Maybe I could've added that repository as a submodule but I chose not to because in this context, the website, I'm really only interested in showing the result.